backend-software/requirements.txt
asgiref==3.2.10
Django==3.1
django-filter==2.4.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
install==1.3.4
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
pandas
gunicorn
